powerpc64le_hello: dump passed FDT
Sort-of dump the passed FDT. We can't dump properties well because
there's no way to tell if a prop is a string, a cell or a blob, so
we dump the size and pointer.

The libfdt bits are from skiboot. string.S is from iquik,
except for memchr, which I had to write just now for libfdt
code to work.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>