rom: more OF CIF
Found a debug option ('-vrdebug', yay for strings!) that produces
lots of fun spew. Implemented peer/child/parent and saw the
DTB being walked.

Fun fact: in OF handle 0 I guess is the first valid phandle and veneer expects
it to have peers. In FDT handle 0 is literally the root and has no peers.
Bit of a hack, I now return no children for 0 and its first subchild as
a peer. Meh.

Hit mfpvr so really need to work on the instruction completion
engine and guest state emulation. Fun.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>