mem: better physical memory tracking
One really shitty available range allocator later, we can
now be sure to not trample any of the ranges used by
the firmware or us (e.g. MA 0 exception vecs,
initrd).

This means we can now assign 128MB of RAM to the VM.

After I refactor the range_blah() stuff out of mem.c,
I will reuse this in rom.c to keep track of OF claims
to properly report available memory. Hopefully this
will prove that tracking ranges at byte (and not page)
level was not a mistake.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>