guest: experimenting with PVR values.
This is kind of cool. I run enough of WNT
to be able to tell which CPUs were supported.
I had no idea the 604e was on the supported list,
but much more exciting, it seems to be okay
with a G3! Wow.

NT on a Wallstreet PDQ anyone? Maybe once I get
iQuikLE working...

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>