guest: Several fixups, getting to MMU fault
In an ideal world this would be a few commits....

Problem 0: emulate cntlzw

Problem 1: NT spinning on acquiring mutex, probably something to do with MMU
(as there's tlbie instruction nearby). It turns out that the code was doing
this:
lwarx
mfsprg0
cmp
stcwx

But according to the 601 manual, an exception will clear your reservation.
Which... means the mutex is never acquired. So the correct course of action
is to not emulate the mfsprg instruction in the hypervisor but instead to just
binary patch the instruction. Linux pr_kvm does this.
I haven't yet worked out the prephv MMU magic to get a magic page though, so
I just use 0... which, umm... works well enough but is without question wrong.

Problem 2: emulate spr 4 and 5 for a "real time clock", poorly.

Problem 3: SPR 1008, hid0 (i think i only hit tihs when using 604, not the 601)

.... and now we can get to the point where we get:
[prephv]FIXME: inject exception into guest
[prephv]Data MMU fault at 0x80136e74 (DAR 0xc0300c00, DSISR 0x40000000)
[prephv]Unrecoverable exception 0x300 from 32-bit

Signed-off-by: Stewart Smith <stewart@flamingspork.com>